Remove impossible or incorrect tests
There were lots of tests for game variations which don't follow a
particular pattern.  For instance, there's no way to generally encode
that "majora" and "ocarina" imply a zelda game.  As soon as the next
made up Zelda word comes out the logic would miss it.

All the remaining tests should be fixable with general rules.