The only Dora is the Explorer
All games currently available which have Dora in the title are Dora the
Explorer games.  People often omit her full title, so we take the luxury
of expanding it.

I added a new test (no-match.t) to guarantee that we didn't munge "dora"
when it occurred inside another word like "dorado" or "pandora".  This
test will be useful if the code is ever refactored.