Treat "spyro" as an abbreviation
There is a series of games formally called The Legend of Spyro.  There
are also other Spyro games.  The phrase "The Legend of" is sometimes
erroneously included or excluded on these titles.  Since "spyro" is such
an unusual name, we can safely treat "spyro" as an abbreviation for the
longer name "the legend of spyro".  This allows us to consistently match
Spyro titles.